:ghost: addon adapter using binding richClient.
