+++ addon/adapter.go
@@ -33,9 +33,36 @@ func init() {
 }
 
 //
+// Client
+type Client = binding.Client
+type Params = binding.Params
+type Param = binding.Param
+type Path = binding.Path
+type Field = binding.Field
+
+//
+// Error
+type SoftError = binding.SoftError
+type ResetError = binding.RestError
+type Conflict = binding.Conflict
+type NotFound = binding.NotFound
+
+//
+// Handler
+type Application = binding.Application
+type Bucket = binding.Bucket
+type BucketContent = binding.BucketContent
+type File = binding.File
+type Identity = binding.Identity
+type Proxy = binding.Proxy
+type RuleSet = binding.RuleSet
+type Setting = binding.Setting
+type Tag = binding.Tag
+type TagCategory = binding.TagCategory
+
+//
 // The Adapter provides hub/addon integration.
 type Adapter struct {
-	// Task API.
 	Task
 	// Settings API
 	Setting Setting
@@ -50,9 +77,9 @@ type Adapter struct {
 	// Tag API.
 	Tag Tag
 	// File API.
-	File binding.File
+	File File
 	// RuleSet API
-	RuleSet binding.RuleSet
+	RuleSet RuleSet
 	// client A REST client.
 	client *Client
 }
@@ -79,11 +106,9 @@ func (h *Adapter) Run(addon func() error) {
 		if err != nil {
 			if _, soft := err.(interface{ Soft() *SoftError }); !soft {
 				Log.Error(err, "Addon failed.")
+				os.Exit(1)
 			}
-			if h.client.Error == nil {
-				h.Failed(err.Error())
-			}
-			os.Exit(1)
+			h.Failed(err.Error())
 		}
 	}()
 	//
@@ -106,48 +131,23 @@ func (h *Adapter) Run(addon func() error) {
 }
 
 //
-// Client provides the REST client.
-func (h *Adapter) Client() *Client {
-	return h.client
-}
-
-//
 // newAdapter builds a new Addon Adapter object.
 func newAdapter() (adapter *Adapter) {
-	//
-	// Build REST client.
-	client := binding.NewClient(Settings.Addon.Hub.URL, Settings.Addon.Hub.Token)
-	//
-	// Build Adapter.
+	richClient := binding.New(Settings.Addon.Hub.URL)
+	richClient.Client.SetToken(Settings.Addon.Hub.Token)
 	adapter = &Adapter{
+		client: richClient.Client,
 		Task: Task{
-			client: client,
+			richClient: richClient,
 		},
-		Setting: Setting{
-			client: client,
-		},
-		Application: Application{
-			client: client,
-		},
-		Identity: Identity{
-			client: client,
-		},
-		Proxy: Proxy{
-			client: client,
-		},
-		TagCategory: TagCategory{
-			client: client,
-		},
-		Tag: Tag{
-			client: client,
-		},
-		File: binding.File{
-			Client: client,
-		},
-		RuleSet: binding.RuleSet{
-			Client: client,
-		},
-		client: client,
+		Setting:     richClient.Setting,
+		Application: richClient.Application,
+		Identity:    richClient.Identity,
+		Proxy:       richClient.Proxy,
+		TagCategory: richClient.TagCategory,
+		Tag:         richClient.Tag,
+		File:        richClient.File,
+		RuleSet:     richClient.RuleSet,
 	}
 
 	Log.Info("Addon (adapter) created.")
