--- addon/client.go
@@ -1,9 +1,0 @@
-package addon
-
-import "github.com/konveyor/tackle2-hub/binding"
-
-type Client = binding.Client
-type Params = binding.Params
-type Param = binding.Param
-type Path = binding.Path
-type Field = binding.Field
