+++ binding/bucket.go
@@ -1,36 +1,89 @@
 package binding
 
-import pathlib "path"
+import (
+	"github.com/konveyor/tackle2-hub/api"
+	pathlib "path"
+)
 
 //
 // Bucket API.
 type Bucket struct {
-	// hub API client.
-	Client *Client
-	// root path
-	path string
+	client *Client
+}
+
+//
+// Create a Bucket.
+func (h *Bucket) Create(r *api.Bucket) (err error) {
+	err = h.client.Post(api.BucketsRoot, &r)
+	return
+}
+
+//
+// Get a bucket.
+func (h *Bucket) Get(id uint) (r *api.Bucket, err error) {
+	r = &api.Bucket{}
+	path := Path(api.BucketRoot).Inject(Params{api.ID: id})
+	err = h.client.Get(path, r)
+	return
+}
+
+//
+// List buckets.
+func (h *Bucket) List() (list []api.Bucket, err error) {
+	list = []api.Bucket{}
+	err = h.client.Get(api.BucketsRoot, &list)
+	return
+}
+
+//
+// Delete a bucket.
+func (h *Bucket) Delete(id uint) (err error) {
+	err = h.client.Delete(Path(api.BucketRoot).Inject(Params{api.ID: id}))
+	return
+}
+
+//
+// Content returns content API.
+func (h *Bucket) Content(id uint) (b *BucketContent) {
+	params := Params{
+		api.Wildcard: "",
+		api.ID:       id,
+	}
+	path := Path(api.BucketRoot).Inject(params)
+	b = &BucketContent{
+		root:   path,
+		client: h.client,
+	}
+	return
+}
+
+//
+// BucketContent API.
+type BucketContent struct {
+	client *Client
+	root   string
 }
 
 //
 // Get reads from the bucket.
-// The source (path) is relative to the bucket root.
-func (h *Bucket) Get(source, destination string) (err error) {
-	err = h.Client.BucketGet(pathlib.Join(h.path, source), destination)
+// The source (root) is relative to the bucket root.
+func (h *BucketContent) Get(source, destination string) (err error) {
+	err = h.client.BucketGet(pathlib.Join(h.root, source), destination)
 	return
 }
 
 //
 // Put writes to the bucket.
-// The destination (path) is relative to the bucket root.
-func (h *Bucket) Put(source, destination string) (err error) {
-	err = h.Client.BucketPut(source, pathlib.Join(h.path, destination))
+// The destination (root) is relative to the bucket root.
+func (h *BucketContent) Put(source, destination string) (err error) {
+	err = h.client.BucketPut(source, pathlib.Join(h.root, destination))
 	return
 }
 
 //
-// Delete deletes content at the specified path.
-// The path is relative to the bucket root.
-func (h *Bucket) Delete(path string) (err error) {
-	err = h.Client.Delete(pathlib.Join(h.path, path))
+// Delete deletes content at the specified root.
+// The source is relative to the bucket root.
+func (h *BucketContent) Delete(path string) (err error) {
+	err = h.client.Delete(pathlib.Join(h.root, path))
 	return
 }
