+++ binding/file.go
@@ -8,21 +8,20 @@ import (
 //
 // File API.
 type File struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Get downloads a file.
 func (h *File) Get(id uint, destination string) (err error) {
 	path := Path(api.FileRoot).Inject(Params{api.ID: id})
-	isDir, err := h.Client.IsDir(destination, false)
+	isDir, err := h.client.IsDir(destination, false)
 	if err != nil {
 		return
 	}
 	if isDir {
 		r := &api.File{}
-		err = h.Client.Get(path, r)
+		err = h.client.Get(path, r)
 		if err != nil {
 			return
 		}
@@ -30,7 +29,7 @@ func (h *File) Get(id uint, destination string) (err error) {
 			destination,
 			r.Name)
 	}
-	err = h.Client.FileGet(path, destination)
+	err = h.client.FileGet(path, destination)
 	return
 }
 
@@ -39,7 +38,7 @@ func (h *File) Get(id uint, destination string) (err error) {
 func (h *File) Put(source string) (r *api.File, err error) {
 	r = &api.File{}
 	path := Path(api.FileRoot).Inject(Params{api.ID: pathlib.Base(source)})
-	err = h.Client.FilePut(path, source, r)
+	err = h.client.FilePut(path, source, r)
 	return
 }
 
@@ -47,6 +46,6 @@ func (h *File) Put(source string) (r *api.File, err error) {
 // Delete a file.
 func (h *File) Delete(id uint) (err error) {
 	path := Path(api.FileRoot).Inject(Params{api.ID: id})
-	err = h.Client.Delete(path)
+	err = h.client.Delete(path)
 	return
 }
