+++ binding/identity.go
@@ -7,14 +7,13 @@ import (
 //
 // Identity API.
 type Identity struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Create a Identity.
 func (h *Identity) Create(r *api.Identity) (err error) {
-	err = h.Client.Post(api.IdentitiesRoot, &r)
+	err = h.client.Post(api.IdentitiesRoot, &r)
 	return
 }
 
@@ -22,8 +21,17 @@ func (h *Identity) Create(r *api.Identity) (err error) {
 // Get a Identity by ID.
 func (h *Identity) Get(id uint) (r *api.Identity, err error) {
 	r = &api.Identity{}
+	p := Param{
+		Key:   api.Decrypted,
+		Value: "1",
+	}
 	path := Path(api.IdentityRoot).Inject(Params{api.ID: id})
-	err = h.Client.Get(path, r)
+	err = h.client.Get(path, r, p)
+	if err != nil {
+		return
+	}
+	m := r.Model()
+	r.With(m)
 	return
 }
 
@@ -31,7 +39,19 @@ func (h *Identity) Get(id uint) (r *api.Identity, err error) {
 // List Identities.
 func (h *Identity) List() (list []api.Identity, err error) {
 	list = []api.Identity{}
-	err = h.Client.Get(api.IdentitiesRoot, &list)
+	p := Param{
+		Key:   api.Decrypted,
+		Value: "1",
+	}
+	err = h.client.Get(api.IdentitiesRoot, &list, p)
+	if err != nil {
+		return
+	}
+	for i := range list {
+		r := &list[i]
+		m := r.Model()
+		r.With(m)
+	}
 	return
 }
 
@@ -39,13 +59,13 @@ func (h *Identity) List() (list []api.Identity, err error) {
 // Update a Identity.
 func (h *Identity) Update(r *api.Identity) (err error) {
 	path := Path(api.IdentityRoot).Inject(Params{api.ID: r.ID})
-	err = h.Client.Put(path, r)
+	err = h.client.Put(path, r)
 	return
 }
 
 //
 // Delete a Identity.
 func (h *Identity) Delete(id uint) (err error) {
-	err = h.Client.Delete(Path(api.IdentityRoot).Inject(Params{api.ID: id}))
+	err = h.client.Delete(Path(api.IdentityRoot).Inject(Params{api.ID: id}))
 	return
 }
