+++ binding/jobfunction.go
@@ -7,14 +7,13 @@ import (
 //
 // JobFunction API.
 type JobFunction struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Create a JobFunction.
 func (h *JobFunction) Create(r *api.JobFunction) (err error) {
-	err = h.Client.Post(api.JobFunctionsRoot, &r)
+	err = h.client.Post(api.JobFunctionsRoot, &r)
 	return
 }
 
@@ -23,7 +22,7 @@ func (h *JobFunction) Create(r *api.JobFunction) (err error) {
 func (h *JobFunction) Get(id uint) (r *api.JobFunction, err error) {
 	r = &api.JobFunction{}
 	path := Path(api.JobFunctionRoot).Inject(Params{api.ID: id})
-	err = h.Client.Get(path, r)
+	err = h.client.Get(path, r)
 	return
 }
 
@@ -31,7 +30,7 @@ func (h *JobFunction) Get(id uint) (r *api.JobFunction, err error) {
 // List JobFunctions.
 func (h *JobFunction) List() (list []api.JobFunction, err error) {
 	list = []api.JobFunction{}
-	err = h.Client.Get(api.JobFunctionsRoot, &list)
+	err = h.client.Get(api.JobFunctionsRoot, &list)
 	return
 }
 
@@ -39,13 +38,13 @@ func (h *JobFunction) List() (list []api.JobFunction, err error) {
 // Update a JobFunction.
 func (h *JobFunction) Update(r *api.JobFunction) (err error) {
 	path := Path(api.JobFunctionRoot).Inject(Params{api.ID: r.ID})
-	err = h.Client.Put(path, r)
+	err = h.client.Put(path, r)
 	return
 }
 
 //
 // Delete a JobFunction.
 func (h *JobFunction) Delete(id uint) (err error) {
-	err = h.Client.Delete(Path(api.JobFunctionRoot).Inject(Params{api.ID: id}))
+	err = h.client.Delete(Path(api.JobFunctionRoot).Inject(Params{api.ID: id}))
 	return
 }
