+++ binding/proxy.go
@@ -7,14 +7,13 @@ import (
 //
 // Proxy API.
 type Proxy struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Create a Proxy.
 func (h *Proxy) Create(r *api.Proxy) (err error) {
-	err = h.Client.Post(api.ProxiesRoot, &r)
+	err = h.client.Post(api.ProxiesRoot, &r)
 	return
 }
 
@@ -23,7 +22,7 @@ func (h *Proxy) Create(r *api.Proxy) (err error) {
 func (h *Proxy) Get(id uint) (r *api.Proxy, err error) {
 	r = &api.Proxy{}
 	path := Path(api.ProxyRoot).Inject(Params{api.ID: id})
-	err = h.Client.Get(path, r)
+	err = h.client.Get(path, r)
 	return
 }
 
@@ -31,7 +30,7 @@ func (h *Proxy) Get(id uint) (r *api.Proxy, err error) {
 // List Proxies.
 func (h *Proxy) List() (list []api.Proxy, err error) {
 	list = []api.Proxy{}
-	err = h.Client.Get(api.ProxiesRoot, &list)
+	err = h.client.Get(api.ProxiesRoot, &list)
 	return
 }
 
@@ -39,13 +38,31 @@ func (h *Proxy) List() (list []api.Proxy, err error) {
 // Update a Proxy.
 func (h *Proxy) Update(r *api.Proxy) (err error) {
 	path := Path(api.ProxyRoot).Inject(Params{api.ID: r.ID})
-	err = h.Client.Put(path, r)
+	err = h.client.Put(path, r)
 	return
 }
 
 //
 // Delete a Proxy.
 func (h *Proxy) Delete(id uint) (err error) {
-	err = h.Client.Delete(Path(api.ProxyRoot).Inject(Params{api.ID: id}))
+	err = h.client.Delete(Path(api.ProxyRoot).Inject(Params{api.ID: id}))
+	return
+}
+
+//
+// Find by Kind.
+// Returns nil when not found.
+func (h *Proxy) Find(kind string) (r *api.Proxy, err error) {
+	list, err := h.List()
+	if err != nil {
+		return
+	}
+	for i := range list {
+		p := &list[i]
+		if p.Kind == kind {
+			r = p
+			break
+		}
+	}
 	return
 }
