+++ binding/ruleset.go
@@ -7,14 +7,13 @@ import (
 //
 // RuleSet API.
 type RuleSet struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Create a RuleSet.
 func (h *RuleSet) Create(r *api.RuleSet) (err error) {
-	err = h.Client.Post(api.RuleSetsRoot, &r)
+	err = h.client.Post(api.RuleSetsRoot, &r)
 	return
 }
 
@@ -23,7 +22,7 @@ func (h *RuleSet) Create(r *api.RuleSet) (err error) {
 func (h *RuleSet) Get(id uint) (r *api.RuleSet, err error) {
 	r = &api.RuleSet{}
 	path := Path(api.RuleSetRoot).Inject(Params{api.ID: id})
-	err = h.Client.Get(path, r)
+	err = h.client.Get(path, r)
 	return
 }
 
@@ -31,7 +30,7 @@ func (h *RuleSet) Get(id uint) (r *api.RuleSet, err error) {
 // List RuleSets.
 func (h *RuleSet) List() (list []api.RuleSet, err error) {
 	list = []api.RuleSet{}
-	err = h.Client.Get(api.RuleSetsRoot, &list)
+	err = h.client.Get(api.RuleSetsRoot, &list)
 	return
 }
 
@@ -39,13 +38,13 @@ func (h *RuleSet) List() (list []api.RuleSet, err error) {
 // Update a RuleSet.
 func (h *RuleSet) Update(r *api.RuleSet) (err error) {
 	path := Path(api.RuleSetRoot).Inject(Params{api.ID: r.ID})
-	err = h.Client.Put(path, r)
+	err = h.client.Put(path, r)
 	return
 }
 
 //
 // Delete a RuleSet.
 func (h *RuleSet) Delete(id uint) (err error) {
-	err = h.Client.Delete(Path(api.RuleSetRoot).Inject(Params{api.ID: id}))
+	err = h.client.Delete(Path(api.RuleSetRoot).Inject(Params{api.ID: id}))
 	return
 }
