+++ binding/stakeholder.go
@@ -7,14 +7,13 @@ import (
 //
 // Stakeholder API.
 type Stakeholder struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Create a Stakeholder.
 func (h *Stakeholder) Create(r *api.Stakeholder) (err error) {
-	err = h.Client.Post(api.StakeholdersRoot, &r)
+	err = h.client.Post(api.StakeholdersRoot, &r)
 	return
 }
 
@@ -23,7 +22,7 @@ func (h *Stakeholder) Create(r *api.Stakeholder) (err error) {
 func (h *Stakeholder) Get(id uint) (r *api.Stakeholder, err error) {
 	r = &api.Stakeholder{}
 	path := Path(api.StakeholderRoot).Inject(Params{api.ID: id})
-	err = h.Client.Get(path, r)
+	err = h.client.Get(path, r)
 	return
 }
 
@@ -31,7 +30,7 @@ func (h *Stakeholder) Get(id uint) (r *api.Stakeholder, err error) {
 // List Stakeholders.
 func (h *Stakeholder) List() (list []api.Stakeholder, err error) {
 	list = []api.Stakeholder{}
-	err = h.Client.Get(api.StakeholdersRoot, &list)
+	err = h.client.Get(api.StakeholdersRoot, &list)
 	return
 }
 
@@ -39,13 +38,13 @@ func (h *Stakeholder) List() (list []api.Stakeholder, err error) {
 // Update a Stakeholder.
 func (h *Stakeholder) Update(r *api.Stakeholder) (err error) {
 	path := Path(api.StakeholderRoot).Inject(Params{api.ID: r.ID})
-	err = h.Client.Put(path, r)
+	err = h.client.Put(path, r)
 	return
 }
 
 //
 // Delete a Stakeholder.
 func (h *Stakeholder) Delete(id uint) (err error) {
-	err = h.Client.Delete(Path(api.StakeholderRoot).Inject(Params{api.ID: id}))
+	err = h.client.Delete(Path(api.StakeholderRoot).Inject(Params{api.ID: id}))
 	return
 }
