+++ binding/stakeholdergroup.go
@@ -7,14 +7,13 @@ import (
 //
 // StakeholderGroup API.
 type StakeholderGroup struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Create a StakeholderGroup.
 func (h *StakeholderGroup) Create(r *api.StakeholderGroup) (err error) {
-	err = h.Client.Post(api.StakeholderGroupsRoot, &r)
+	err = h.client.Post(api.StakeholderGroupsRoot, &r)
 	return
 }
 
@@ -23,7 +22,7 @@ func (h *StakeholderGroup) Create(r *api.StakeholderGroup) (err error) {
 func (h *StakeholderGroup) Get(id uint) (r *api.StakeholderGroup, err error) {
 	r = &api.StakeholderGroup{}
 	path := Path(api.StakeholderGroupRoot).Inject(Params{api.ID: id})
-	err = h.Client.Get(path, r)
+	err = h.client.Get(path, r)
 	return
 }
 
@@ -31,7 +30,7 @@ func (h *StakeholderGroup) Get(id uint) (r *api.StakeholderGroup, err error) {
 // List StakeholderGroups.
 func (h *StakeholderGroup) List() (list []api.StakeholderGroup, err error) {
 	list = []api.StakeholderGroup{}
-	err = h.Client.Get(api.StakeholderGroupsRoot, &list)
+	err = h.client.Get(api.StakeholderGroupsRoot, &list)
 	return
 }
 
@@ -39,13 +38,13 @@ func (h *StakeholderGroup) List() (list []api.StakeholderGroup, err error) {
 // Update a StakeholderGroup.
 func (h *StakeholderGroup) Update(r *api.StakeholderGroup) (err error) {
 	path := Path(api.StakeholderGroupRoot).Inject(Params{api.ID: r.ID})
-	err = h.Client.Put(path, r)
+	err = h.client.Put(path, r)
 	return
 }
 
 //
 // Delete a StakeholderGroup.
 func (h *StakeholderGroup) Delete(id uint) (err error) {
-	err = h.Client.Delete(Path(api.StakeholderGroupRoot).Inject(Params{api.ID: id}))
+	err = h.client.Delete(Path(api.StakeholderGroupRoot).Inject(Params{api.ID: id}))
 	return
 }
