+++ binding/tag.go
@@ -7,14 +7,13 @@ import (
 //
 // Tag API.
 type Tag struct {
-	// hub API client.
-	Client *Client
+	client *Client
 }
 
 //
 // Create a Tag.
 func (h *Tag) Create(r *api.Tag) (err error) {
-	err = h.Client.Post(api.TagsRoot, &r)
+	err = h.client.Post(api.TagsRoot, &r)
 	return
 }
 
@@ -23,7 +22,7 @@ func (h *Tag) Create(r *api.Tag) (err error) {
 func (h *Tag) Get(id uint) (r *api.Tag, err error) {
 	r = &api.Tag{}
 	path := Path(api.TagRoot).Inject(Params{api.ID: id})
-	err = h.Client.Get(path, r)
+	err = h.client.Get(path, r)
 	return
 }
 
@@ -31,7 +30,7 @@ func (h *Tag) Get(id uint) (r *api.Tag, err error) {
 // List Tags.
 func (h *Tag) List() (list []api.Tag, err error) {
 	list = []api.Tag{}
-	err = h.Client.Get(api.TagsRoot, &list)
+	err = h.client.Get(api.TagsRoot, &list)
 	return
 }
 
@@ -39,13 +38,50 @@ func (h *Tag) List() (list []api.Tag, err error) {
 // Update a Tag.
 func (h *Tag) Update(r *api.Tag) (err error) {
 	path := Path(api.TagRoot).Inject(Params{api.ID: r.ID})
-	err = h.Client.Put(path, r)
+	err = h.client.Put(path, r)
 	return
 }
 
 //
 // Delete a Tag.
 func (h *Tag) Delete(id uint) (err error) {
-	err = h.Client.Delete(Path(api.TagRoot).Inject(Params{api.ID: id}))
+	err = h.client.Delete(Path(api.TagRoot).Inject(Params{api.ID: id}))
+	return
+}
+
+//
+// Find by name and type.
+func (h *Tag) Find(name string, category uint) (r *api.Tag, found bool, err error) {
+	list := []api.Tag{}
+	path := Path(api.TagCategoryTagsRoot).Inject(Params{api.ID: category})
+	err = h.client.Get(
+		path,
+		&list,
+		Param{
+			Key:   api.Name,
+			Value: name,
+		})
+	if err != nil {
+		return
+	}
+	if len(list) > 0 {
+		found = true
+		r = &list[0]
+	}
+	return
+}
+
+//
+// Ensure a tag exists.
+func (h *Tag) Ensure(wanted *api.Tag) (err error) {
+	tag, found, err := h.Find(wanted.Name, wanted.Category.ID)
+	if err != nil {
+		return
+	}
+	if !found {
+		err = h.Create(wanted)
+	} else {
+		*wanted = *tag
+	}
 	return
 }
